refactor(anni): reorganize common components
