+++ annotation-interface/components/common/indicators/LoadingSpinner.tsx
