+++ annotation-interface/components/common/interaction/BackButton.tsx
@@ -1,7 +1,7 @@
 import { ChevronLeftIcon } from '@heroicons/react/outline';
 import { useRouter } from 'next/router';
 
-import WithIcon from './WithIcon';
+import WithIcon from '../WithIcon';
 
 export const BackButton = () => {
     const router = useRouter();
