+++ annotation-interface/components/common/interaction/DisplayLanguagePicker.tsx
@@ -1,5 +1,5 @@
-import { supportedLanguages } from '../../common/definitions';
-import { useLanguageContext } from '../../contexts/language';
+import { supportedLanguages } from '../../../common/definitions';
+import { useLanguageContext } from '../../../contexts/language';
 import SelectionPopover from './SelectionPopover';
 
 const DisplayLanguagePicker = () => {
