+++ annotation-interface/components/common/interaction/SearchBar.tsx
